Simplify HermiteSimpsonSolver.cs by extracting objective function computation
Extract objective function computation logic to ObjectiveFunctionFactory.cs,
reducing HermiteSimpsonSolver.cs from 728 to 630 lines (~100 lines removed).

This follows the previous extraction of constraint configuration to
ConstraintConfigurator.cs. The new ObjectiveFunctionFactory contains:
- ComputeTotalCost: Computes running + terminal cost
- ComputeObjectiveGradient: Computes objective gradient using analytical or numerical methods
- CanUseAnalyticalGradientsForCosts: Checks if analytical gradients are available

Closes #12
